open refine

+++ dataprep/views.py
@@ -1,6 +1,8 @@
-from django.shortcuts import render,redirect
+from django.http import HttpResponseRedirect
+from django.shortcuts import render, redirect
+from django.urls import reverse
+
 from .forms import DataPrepForm
-from django.http import HttpResponseRedirect, HttpResponse
 from django.conf import settings
 from .models import DataPrep
 from .dataPrepScript import runDataPrepForForm
@@ -9,12 +11,10 @@ from pathlib import Path
 
 # Create your views here.
 def dataprep(request):
-    submitted = False
-
     if request.method == "POST":
         form = DataPrepForm(request.POST, request.FILES)
         faculties = DataPrep()
-        message = ''
+
         if form.is_valid():
             faculties.fieldQnum = form.cleaned_data['fieldQnum']
             faculties.departmentQnum = form.cleaned_data['departmentQnum']
@@ -23,14 +23,14 @@ def dataprep(request):
             faculties.file = form.cleaned_data['file']
             faculties.save()
 
-            message = runDataPrepForForm(settings.MEDIA_ROOT+'/', str(faculties.file), faculties.fieldQnum,
+            respond = runDataPrepForForm(settings.MEDIA_ROOT+'/', str(faculties.file), faculties.fieldQnum,
                                          faculties.departmentQnum, faculties.websiteBaseUrl,
                                          faculties.allOnePage)
-            # if ('error' in message) or ('No file' in message):
-            [f.unlink() for f in Path(settings.MEDIA_ROOT+'/').glob("*") if f.is_file()]
 
-        submitted = True
-        return HttpResponse(message)
+            if ('error' in respond['message']) or ('No file' in respond['message']):
+                [f.unlink() for f in Path(settings.MEDIA_ROOT+'/').glob("*") if f.is_file()]
+            print("Going to openrefine.")
+            return render(request, 'dataprep/dataPrepForm.html', context={'contextData': respond})
     else:
         form = DataPrepForm
